+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 * Enhancements
   * Support `before_send/1` callbacks
-  * Add Plug.Static
+  * Add `Plug.Static`
   * Allow iodata as the body
   * Do not allow response headers to be set if the response was already sent
+  * Add `Plug.Conn.private` to be used as storage by libraries/frameworks
 
 # v0.3.0